feat: move structs to sepearte module
